modify test to not depend on internal rmarkdown structure
This test was probing the `rmarkdown_output_format` for a particular length and particular list element names. This is an internal rmarkdown data structure so can and does change. In the most recent version we added an element to this list which broke this test. This patch removes the test for specific length and list elements and just tests that the returned element is a named object of the correct class.